+++ tests/test_pipdeptree.py
@@ -1,9 +1,9 @@
 import platform
 import subprocess
 import sys
-import textwrap
 from contextlib import contextmanager
 from itertools import chain
+from pathlib import Path
 from tempfile import NamedTemporaryFile
 
 try:
@@ -493,22 +493,6 @@ def test_custom_interpreter(tmp_path, monkeypatch, capfd, args_joined):
 
 
 def test_guess_version_setuptools():
-    code = """
-    import sys
-
-    import pipdeptree
-
-    if sys.version_info >= (3, 8):
-        import importlib.metadata as importlib_metadata
-    else:
-        import importlib_metadata
-
-    def raise_import_error(name):
-        raise ImportError(name)
-
-    importlib_metadata.version = raise_import_error
-
-    print(pipdeptree.guess_version("setuptools"), end="")
-    """
-    p = subprocess.run([sys.executable, "-c", textwrap.dedent(code)], capture_output=True, check=True)
-    assert p.stdout == b"?"
+    script = Path(__file__).parent / "guess_version_setuptools.py"
+    output = subprocess.check_output([sys.executable, script], text=True)
+    assert output == "?"
